Create deleteme.html
On day 3 we learnt about git that what is git and for what purpose we use git. We also learnt about the basic command of git and get knowledge about how we work on git.
On days 4 and 5 we learnt about MySQL, On day 4 we basically looked in the theory part of the subject and on day 5th we moved towards the query language used in MySQL i.e SQL and looked into some basic queries. This is all about days 3rd, 4th and 5th. We will be back next weekend to learn more about MySQL.